Windows fixes
1) 'true' is not a valid command in Window's cmd.exe
2) Avoid a crash on exit due to the ConsoleThread being killed
3) Fix warning in int -> void * conversion on 64bit builds
4) Win10 1803 now comes with a bsdtar in c:\windows that actually works,
   but it returns windows-style \r\n line breaks

+++ src/launcher-lib/ipcwrapperobject.cpp
@@ -306,7 +306,7 @@ int IpcWrapperObject::qt_metacall(QMetaObject::Call _c, int _id, void **_a)
             value = QVariant::fromValue(dbv);
         }
         // the boolean 'was-successful' return code is stored in _a[1]
-        _a[1] = reinterpret_cast<void *>(mp.write(m_dbusInterface, value) ? 1 : 0);
+        _a[1] = reinterpret_cast<void *>(mp.write(m_dbusInterface, value) ? intptr_t(1) : intptr_t(0));
         break;
     }
     case QMetaObject::InvokeMetaMethod: {
